Remove published files when destroyed or closed

+++ app/models/gp_category/category.rb
@@ -51,6 +51,7 @@ class GpCategory::Category < ActiveRecord::Base
 
   after_save :move_published_files
   after_save :publish_ancestor_pages
+  after_save :clean_published_files
   after_destroy :clean_published_files
   after_destroy :publish_ancestor_pages
 
@@ -205,6 +206,7 @@ class GpCategory::Category < ActiveRecord::Base
   end
 
   def clean_published_files
+    return if !destroyed? && public?
     FileUtils.rm_r(public_path) if public_path.present? && ::File.exist?(public_path)
     FileUtils.rm_r(public_smart_phone_path) if public_smart_phone_path.present? && ::File.exist?(public_smart_phone_path)
   end
