+++ app/models/gp_category/category_type.rb
@@ -36,6 +36,7 @@ class GpCategory::CategoryType < ActiveRecord::Base
   scope :public, -> { where(state: 'public') }
   scope :none, -> { where("#{self.table_name}.id IS ?", nil).where("#{self.table_name}.id IS NOT ?", nil) }
 
+  after_save :clean_published_files
   after_destroy :clean_published_files
 
   def public_categories
@@ -142,6 +143,7 @@ class GpCategory::CategoryType < ActiveRecord::Base
   end
 
   def clean_published_files
+    return if !destroyed? && public?
     FileUtils.rm_r(public_path) if public_path.present? && ::File.exist?(public_path)
     FileUtils.rm_r(public_smart_phone_path) if public_smart_phone_path.present? && ::File.exist?(public_smart_phone_path)
   end
